Struts 2 7.0.x 업그레이드 - 시작 (#11)
라이브러리 버전업과, JUnit 5 커스텀 플러그인만 수정/테스트만 해보았다.
