+++ plugins/struts2-junit5-user-custom-plugin/src/main/java/org/apache/struts2/junit/StrutsSpringJUnit5TestCase.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2;
+package org.apache.struts2.junit;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
