+++ plugins/struts2-junit5-user-custom-plugin/src/test/java/org/apache/struts2/junit/session/SessionSetAction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2.session;
+package org.apache.struts2.junit.session;
 
 import org.apache.struts2.ActionContext;
 import org.apache.struts2.ActionSupport;
