+++ plugins/struts2-junit5-user-custom-plugin/src/test/resources/applicationContext.xml
@@ -23,16 +23,16 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean class="org.springframework.beans.factory.config.CustomScopeConfigurer">
-        <property name="scopes">
-            <map>
-                <entry key="session">
-                    <bean class="org.springframework.context.support.SimpleThreadScope"/>
-                </entry>
-            </map>
-        </property>
-    </bean>
+  <bean class="org.springframework.beans.factory.config.CustomScopeConfigurer">
+    <property name="scopes">
+      <map>
+        <entry key="session">
+          <bean class="org.springframework.context.support.SimpleThreadScope"/>
+        </entry>
+      </map>
+    </property>
+  </bean>
 
-    <bean id="testAction" class="org.apache.struts2.JUnitTestAction"/>
-    <bean id="mySessionBean" class="org.apache.struts2.MySessionBean" scope="session" />
+  <bean id="testAction" class="org.apache.struts2.junit.JUnitTestAction"/>
+  <bean id="mySessionBean" class="org.apache.struts2.junit.MySessionBean" scope="session"/>
 </beans>
