+++ plugins/struts2-junit5-user-custom-plugin/src/test/resources/struts-session-values-test.xml
@@ -26,10 +26,10 @@
 
 <struts>
     <package name="test" namespace="/sessiontest" extends="struts-default">
-        <action name="sessionSet" class="org.apache.struts2.session.SessionSetAction">
+        <action name="sessionSet" class="org.apache.struts2.junit.session.SessionSetAction">
             <result type="freemarker">/template-session.ftl</result>
         </action>
-        <action name="sessionGet" class="org.apache.struts2.session.SessionGetAction">
+        <action name="sessionGet" class="org.apache.struts2.junit.session.SessionGetAction">
             <result type="freemarker">/template-session.ftl</result>
         </action>
     </package>
