+++ plugins/struts2-junit5-user-custom-plugin/src/test/resources/struts-test.xml
@@ -25,7 +25,7 @@
 
 <struts>
     <package name="test" namespace="/test" extends="struts-default">
-        <action name="testAction-2" class="org.apache.struts2.JUnitTestAction">
+        <action name="testAction-2" class="org.apache.struts2.junit.JUnitTestAction">
             <result type="freemarker">/template-2.ftl</result>
         </action>
     </package>
