+++ plugins/struts2-junit5-user-custom-plugin/src/test/resources/struts.xml
@@ -26,7 +26,7 @@
 <struts>
     <constant name="struts.objectFactory" value="spring"/>
     <package name="test" namespace="/test" extends="struts-default">
-        <action name="testAction" class="org.apache.struts2.JUnitTestAction">
+        <action name="testAction" class="org.apache.struts2.junit.JUnitTestAction">
             <result type="freemarker">/template-1.ftl</result>
         </action>
     </package>
