+++ pom.xml
@@ -9,7 +9,7 @@
   <artifactId>struts2-study</artifactId>
   <version>1.0.0-SNAPSHOT</version>
   <packaging>pom</packaging>
-  
+
   <name>${project.artifactId} - Struts 2 Study Maven Parent</name>
   <url>https://github.com/fp024/struts2-study/struts2-study-parent</url>
 
@@ -20,67 +20,73 @@
     <maven.compiler.target>${java.version}</maven.compiler.target>
 
     <!-- Servlet 버전 -->
+    <!--
     <javax.servlet-api.version>4.0.1</javax.servlet-api.version>
     <javax.servlet.jsp-api.version>2.3.3</javax.servlet.jsp-api.version>
-    
-    
-    <struts2.version>6.7.0</struts2.version>
-    <spring.platformVersion>5.3.39</spring.platformVersion>
+    -->
+
+    <jakarta.servlet-api.version>6.0.0</jakarta.servlet-api.version>
+    <jakarta.servlet.jsp-api.version>3.1.0</jakarta.servlet.jsp-api.version>
+
+
+    <struts2.version>7.0.3</struts2.version>
+    <spring.platformVersion>6.2.5</spring.platformVersion>
     
     <struts2-junit5-user-custom-plugin.version>${project.version}</struts2-junit5-user-custom-plugin.version>
-    
-    
-    <log4j2.version>2.24.2</log4j2.version>
-    
+
+
+    <log4j2.version>2.24.3</log4j2.version>
+
     <!-- Jackson 버전이 뒤섞이는 경우가 있어서 jackson-bom을 추가했다. -->
-    <jackson.version>2.18.2</jackson.version>
-    
+    <jackson.version>2.18.3</jackson.version>
+
     <flexjson.version>3.3</flexjson.version>
 
-    <hikaricp.version>6.2.1</hikaricp.version>
-    <hibernate.version>5.6.15.Final</hibernate.version>
+    <hikaricp.version>6.3.0</hikaricp.version>
+    <hibernate.version>6.6.13.Final</hibernate.version>
     <hsqldb.version>2.7.4</hsqldb.version>
 
-    <javax.annotation-api.version>1.3.2</javax.annotation-api.version>
+    <!-- <javax.annotation-api.version>1.3.2</javax.annotation-api.version> -->
+    <javax.annotation-api.version>2.1.1</javax.annotation-api.version>
 
     <jasperreports.version>6.21.4</jasperreports.version>
 
-    <lombok.version>1.18.36</lombok.version>
+    <lombok.version>1.18.38</lombok.version>
 
-    <junit.version>5.11.3</junit.version>
-    <mockito.version>5.14.2</mockito.version>
+    <junit.version>5.12.1</junit.version>
+    <mockito.version>5.17.9</mockito.version>
     <hamcrest.version>3.0</hamcrest.version>
-    <assertj.version>3.26.3</assertj.version>
+    <assertj.version>3.27.3</assertj.version>
 
     <!-- struts2-convention-plugin에서 Java 17 바이트 코드 해석을 못해서, 버전을 올렸다. -->
-    <asm.version>9.7.1</asm.version>
+    <asm.version>9.8</asm.version>
 
-    <byte-buddy.version>1.15.10</byte-buddy.version>
+    <byte-buddy.version>1.17.5</byte-buddy.version>
 
-    <jetty.version>12.0.15</jetty.version>
+    <jetty.version>12.0.19</jetty.version>
     <jetty-plugin-group-id>org.eclipse.jetty.ee8</jetty-plugin-group-id>
-    <jetty-plugin-artifact-id>jetty-ee8-maven-plugin</jetty-plugin-artifact-id>
+    <jetty-plugin-artifact-id>jetty-ee10-maven-plugin</jetty-plugin-artifact-id>
     
-    <jetty-root-jstl-group-id>org.eclipse.jetty.ee8</jetty-root-jstl-group-id>
-    <jetty-root-jstl-artifact-id>jetty-ee8-glassfish-jstl</jetty-root-jstl-artifact-id>
+    <jetty-root-jstl-group-id>org.eclipse.jetty.ee10</jetty-root-jstl-group-id>
+    <jetty-root-jstl-artifact-id>jetty-ee10-glassfish-jstl</jetty-root-jstl-artifact-id>
     
     <jetty-exclude-jstl-group-id>org.glassfish.web</jetty-exclude-jstl-group-id>
-    <jetty-exclude-jstl-artifact-id>javax.servlet.jsp.jstl</jetty-exclude-jstl-artifact-id>
-    <!-- https://mvnrepository.com/artifact/org.glassfish.web/javax.servlet.jsp.jstl -->
-    <jetty-exclude-jstl-version>1.2.5</jetty-exclude-jstl-version>
+    <jetty-exclude-jstl-artifact-id>jakarta.servlet.jsp.jstl</jetty-exclude-jstl-artifact-id>
+    <!-- https://mvnrepository.com/artifact/org.glassfish.web/jakarta.servlet.jsp.jstl -->
+    <jetty-exclude-jstl-version>3.0.1</jetty-exclude-jstl-version>
 
 
-    <jetty10.version>10.0.24</jetty10.version>
-    <jetty10-plugin-group-id>org.eclipse.jetty</jetty10-plugin-group-id>
-    <jetty10-plugin-artifact-id>jetty-maven-plugin</jetty10-plugin-artifact-id>
+    <jetty11.version>11.0.25</jetty11.version>
+    <jetty11-plugin-group-id>org.eclipse.jetty</jetty11-plugin-group-id>
+    <jetty11-plugin-artifact-id>jetty-maven-plugin</jetty11-plugin-artifact-id>
     
-    <jetty10-root-jstl-group-id>org.eclipse.jetty</jetty10-root-jstl-group-id>
-    <jetty10-root-jstl-artifact-id>apache-jstl</jetty10-root-jstl-artifact-id>
+    <jetty11-root-jstl-group-id>org.eclipse.jetty</jetty11-root-jstl-group-id>
+    <jetty11-root-jstl-artifact-id>glassfish-jstl</jetty11-root-jstl-artifact-id>
     
-    <jetty10-exclude-jstl-group-id>org.apache.taglibs</jetty10-exclude-jstl-group-id>
-    <jetty10-exclude-jstl-artifact-id>taglibs-standard-impl</jetty10-exclude-jstl-artifact-id>
-    <jetty10-exclude-jstl-version>1.2.5</jetty10-exclude-jstl-version>
-    
+    <jetty11-exclude-jstl-group-id>org.glassfish.web</jetty11-exclude-jstl-group-id>
+    <jetty11-exclude-jstl-artifact-id>jakarta.servlet.jsp.jstl</jetty11-exclude-jstl-artifact-id>
+    <jetty11-exclude-jstl-version>2.0.0</jetty11-exclude-jstl-version>
+
 
     <!-- <jettyContextPath>/${project.build.finalName}</jettyContextPath>-->
     <!-- 예제 프로젝트로서 개별 실행하므로 경로를 ContextPath로 나눌 필요는 없다. -->
@@ -91,16 +97,16 @@
     <jettyAutoDeploy>-1</jettyAutoDeploy>
 
     <!-- Maven Plugin 버전 -->
-    <maven-clean-plugin.version>3.4.0</maven-clean-plugin.version>
+    <maven-clean-plugin.version>3.4.1</maven-clean-plugin.version>
     <maven-resources-plugin.version>3.3.1</maven-resources-plugin.version>
-    <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
-    <maven-surefire-plugin.version>3.5.2</maven-surefire-plugin.version>
-    <maven-failsafe-plugin.version>3.5.2</maven-failsafe-plugin.version>
+    <maven-compiler-plugin.version>3.14.0</maven-compiler-plugin.version>
+    <maven-surefire-plugin.version>3.5.3</maven-surefire-plugin.version>
+    <maven-failsafe-plugin.version>3.5.3</maven-failsafe-plugin.version>
     <maven-war-plugin.version>3.4.0</maven-war-plugin.version>
-    <maven-install-plugin.version>3.1.3</maven-install-plugin.version>
-    <maven-deploy-plugin.version>3.1.3</maven-deploy-plugin.version>
+    <maven-install-plugin.version>3.1.4</maven-install-plugin.version>
+    <maven-deploy-plugin.version>3.1.4</maven-deploy-plugin.version>
     <maven-toolchains-plugin.version>3.2.0</maven-toolchains-plugin.version>
-    
+
     <!-- Struts JUnit Plugin -->
     <commons-compress.version>1.27.1</commons-compress.version>
     <portlet-api.version>3.0.1</portlet-api.version>
