+++ struts2-study-parent/pom.xml
@@ -17,19 +17,18 @@
 
   <profiles>
     <profile>
-      <id>jetty10</id>
+      <id>jetty11</id>
       <properties>
-        <jetty.version>${jetty10.version}</jetty.version>
-        <jetty-plugin-group-id>${jetty10-plugin-group-id}</jetty-plugin-group-id>
-        <jetty-plugin-artifact-id>${jetty10-plugin-artifact-id}</jetty-plugin-artifact-id>
+        <jetty.version>${jetty11.version}</jetty.version>
+        <jetty-plugin-group-id>${jetty11-plugin-group-id}</jetty-plugin-group-id>
+        <jetty-plugin-artifact-id>${jetty11-plugin-artifact-id}</jetty-plugin-artifact-id>
 
-        <jetty-root-jstl-group-id>${jetty10-root-jstl-group-id}</jetty-root-jstl-group-id>
-        <jetty-root-jstl-artifact-id>apache-jstl</jetty-root-jstl-artifact-id>
+        <jetty-root-jstl-group-id>${jetty11-root-jstl-group-id}</jetty-root-jstl-group-id>
+        <jetty-root-jstl-artifact-id>${jetty11-root-jstl-artifact-id}</jetty-root-jstl-artifact-id>
 
-        <jetty-exclude-jstl-group-id>${jetty10-exclude-jstl-group-id}</jetty-exclude-jstl-group-id>
-        <jetty-exclude-jstl-artifact-id>${jetty10-exclude-jstl-artifact-id}
-        </jetty-exclude-jstl-artifact-id>
-        <jetty-exclude-jstl-version>${jetty10-exclude-jstl-version}</jetty-exclude-jstl-version>
+        <jetty-exclude-jstl-group-id>${jetty11-exclude-jstl-group-id}</jetty-exclude-jstl-group-id>
+        <jetty-exclude-jstl-artifact-id>${jetty11-exclude-jstl-artifact-id}</jetty-exclude-jstl-artifact-id>
+        <jetty-exclude-jstl-version>${jetty11-exclude-jstl-version}</jetty-exclude-jstl-version>
       </properties>
     </profile>
   </profiles>
